+++ PluginInstallation/CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/).
 ### Changed
 - Update to latest core v1.3.1
 - Fixed sending full address 
+- Fixed using first and last name for drop-off address from shipping address instead of the customer.
 
 ## [v2.0.4](https://github.com/logeecom/pl_prestashop_module/compare/v2.0.4...v2.0.3) - 2019-07-11
 ### Changed